Refactored geomoptics + waveoptics modules
